requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.10
async-timeout==4.0.2
Babel==2.10.3
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.10.0
Brotli==1.0.9
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.7.0
Deprecated==1.2.13
dill==0.3.6
dj-database-url==1.0.0
Django==4.0.8
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-money==3.0.0
django-phonenumber-field==7.0.0
et-xmlfile==1.1.0
fonttools==4.38.0
gunicorn==20.1.0
html5lib==1.1
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pathspec==0.10.1
phonenumbers==8.13.0
Pillow==9.3.0
platformdirs==2.5.3
prompt-toolkit==3.0.32
psycopg2==2.9.5
psycopg2-binary==2.9.5
py-moneyed==2.0
pycparser==2.21
pydyf==0.5.0
pylint==2.15.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
pypandoc==1.10
pyparsing==3.0.9
pyphen==0.13.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytz==2022.4
redis==4.3.4
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tinycss2==1.1.1
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.4.0
tzdata==2022.4
vine==5.0.0
wcwidth==0.2.5
weasyprint==57.1
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.1
